requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
beautifulsoup4
requests
gTTS
google-generativeai